Move Task.database.functions to Task.controller
